+++ packages/react-filterlist/__stories__/DeferredInit/index.tsx
@@ -161,6 +161,11 @@ const List: FC<RouteComponentProps> = (props) => {
     [filterlist],
   );
 
+  const reload = useCallback(
+    (): Promise<void> => filterlist.reload(),
+    [filterlist],
+  );
+
   const resetFilter = useCallback((
     filterName: string,
   ): Promise<void> => filterlist.resetFilter(
@@ -200,6 +205,7 @@ const List: FC<RouteComponentProps> = (props) => {
       applyFilter={applyFilter}
       setAndApplyFilter={setAndApplyFilter}
       resetAllFilters={resetAllFilters}
+      reload={reload}
       setSorting={setSorting}
     />
   );