+++ packages/react-filterlist/__stories__/InfinityList/index.tsx
@@ -147,6 +147,11 @@ const List: FC<RouteComponentProps> = (props) => {
     [filterlist],
   );
 
+  const reload = useCallback(
+    (): Promise<void> => filterlist.reload(),
+    [filterlist],
+  );
+
   const resetFilter = useCallback((
     filterName: string,
   ): Promise<void> => filterlist.resetFilter(
@@ -171,7 +176,6 @@ const List: FC<RouteComponentProps> = (props) => {
     additional,
     items,
     loading,
-    loadedPages,
 
     sort,
 
@@ -191,6 +195,7 @@ const List: FC<RouteComponentProps> = (props) => {
       applyFilter={applyFilter}
       setAndApplyFilter={setAndApplyFilter}
       resetAllFilters={resetAllFilters}
+      reload={reload}
       setSorting={setSorting}
       isInfinity
       loadMore={loadMore}